All of the css I added to polish
(so I did this after I had worked with it so it’s all my changes as
once) — I found the background color by having the website we used in
class match it - the only issue I came across was in trying to change
the color of the nav-bar I couldn’t use a made color - it would make a
box surrounding the word opposed to changing the word so I instead used
a standard blue. Also the white for the last table doesn’t reach all
the way horizontally.